Feb Branch 2018

+++ src/main/java/com/benRem/brPoMgmt/repository/PurchaseOrderRepository.java
@@ -4,8 +4,10 @@ import com.benRem.brPoMgmt.domain.PurchaseOrder;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface PurchaseOrderRepository extends CrudRepository<PurchaseOrder, Long>{
+public interface PurchaseOrderRepository extends CrudRepository<PurchaseOrder, BigInteger>{
 	//List<Purchase_Order> findAllPurchaseOrder();
